Created main.py
Set up the basis for the magpie script.

Magpie is a script that aims to be a notification center for
internet content, delievered through telegram. Ideally, the script
will fetch updates from various internet services, i.e.
- Facebook: New posts on a page, filterable by keywords
- Twitch: When a channel goes live
- RSS: When a RSS feed updates
- Reddit: When a reddit account receives a PM
- etc.

A basic skeleton was set up with this commit. I was particularly
focused on setting up a json config file. This will:
- Allow users to change list of services for magpie without edits
  to the main script at main.py
- Allow me to easily commit changes to github (instead of to my
  private git remote repo, and then transfering to github) since
  OAuth keys will be in config.json only, which is not tracked by
  git.

The script ends with class web_grabber. This is scrap code written
as a by-product of trying to figure out the process.

TODO:
- Add a history file (csv)
- Add class for telegram api
- Clean, rename, and organise web_grabber
- Add other classes that fetch updates