Modify auto-created config.json to include 'telegram_chanel_id'
This is the channel id which the magpie bot will send updates to.
Though it is possible to request only channel name in config.json,
then run code to send the message,
    {"ok":false,"error_code":401,"description":"Unauthorized"}
to
    chat_id="@channelname"
Doing this will actually send above json coded message to the
channel, producing spam.

TODO: It may be possible to...
- if channel_name in config and channel_id not in config
- bot.sendMessage(query_json)
- config_json['accounts'].append(reply)