Minor change so that empty 'service-name' entries are ignored
Magpie will check if service-name field in config.json for the item
is an empty string. If so, it will ignore that item.

This allows for the inclusion of 'template' items for each service
with empty keys. This template json is generated on first run set-
up of the script.

As a result, the first-run generation of config.json has also been
slightly altered.