Fix issue where magpie_sync could not save files
The argument passed to telegram.File.download relating to the dir
which file is to be saved represents the full actual path including
name, not just path leading to. Fixed this by finding file_path
attribute of telegram.File object and extracting the file_name
from the end of the url.