Add look.py, send_new_update (twitch.py), some minor changes
twitch.__init__ is now called in listen.py at the start of every
command to reload config (should there be changes between commands
from look.py or otherwise, they will be reflected in the instance
variables).

look.py is the script that fetches updates and sends them when a
new update is detected.